requirements.txt
absl-py==0.11.0
aflow==0.0.11
alabaster==0.7.8
anyio==2.0.2
appdirs==1.4.4
appnope==0.1.2
apturl==0.5.2
argon2-cffi==20.1.0
arrow==0.17.0
ase==3.20.1
asn1crypto==0.24.0
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.9.3
binaryornot==0.4.4
bleach==3.2.1
blinker==1.4
Brlapi==0.7.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
citrination-client==6.5.1
click==7.1.2
colorama==0.4.3
command-not-found==0.3
cookiecutter==1.7.2
cryptography==3.3.1
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
dill==0.3.3
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
docutils==0.16
duplicity==0.8.12.0
entrypoints==0.3
fair-research-login==0.2.0
fasteners==0.14.1
filelock==3.0.12
flatbuffers==1.12
future==0.18.2
gast==0.3.3
git-filter-repo==2.29.0
globus-nexus-client==0.2.9
globus-sdk==1.10.0
google-auth==1.26.1
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
graphviz==0.16
grpcio==1.32.0
h5py==2.10.0
httplib2==0.14.0
idna==2.10
imagesize==1.2.0
imbalanced-learn==0.8.0
imblearn==0.0
IProgress==0.4
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jarvis==0.1.9
jarvis-tools==2021.2.3
jedi==0.18.0
Jinja2==2.11.2
jinja2-time==0.2.0
joblib==1.0.0
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyter-kernel-gateway==2.4.3
jupyter-server==1.2.0
jupyterlab==3.0.1
jupyterlab-pygments==0.1.2
jupyterlab-server==2.0.0
jupyterlab-widgets==1.0.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
keyrings.alt==3.4.0
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.3
MarkupSafe==1.1.1
matminer==0.6.4
matplotlib==3.3.3
mdf-forge==0.7.6
mdf-toolbox==0.5.7
mistune==0.8.4
modnet==0.1.8
monotonic==1.5
monty==4.0.2
mpmath==1.1.0
nbclassic==0.2.5
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nbstripout==0.3.7
nest-asyncio==1.4.3
netifaces==0.10.4
networkx==2.5
notebook==6.1.6
numpy==1.19.4
oauth==1.0.1
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
packaging==20.8
palettable==3.3.0
pandas==1.2.0
pandocfilters==1.4.3
paramiko==2.6.0
parso==0.8.1
pathlib==1.0.1
pbr==5.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
Pint==0.16.1
pkginfo==1.7.0
plotly==4.14.1
poyo==0.5.0
prometheus-client==0.9.0
prompt-toolkit==3.0.9
protobuf==3.14.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pydot==1.4.1
Pygments==2.7.3
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymatgen==2020.12.31
pymongo==3.11.2
PyNaCl==1.3.0
pyparsing==2.4.7
pypif==2.1.2
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.17.3
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-dotenv==0.15.0
python-slugify==4.0.1
pytz==2020.5
pyxdg==0.26
PyYAML==5.3.1
pyzmq==20.0.0
qmpy-rester==0.2.0
qtconsole==5.0.2
QtPy==1.9.0
readme-renderer==28.0
reportlab==3.5.34
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
retrying==1.3.3
rfc3986==1.4.0
roman==2.0.0
rsa==4.7
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scikit-learn==0.24.0
scikit-plot==0.3.7
scipy==1.6.0
seaborn==0.11.1
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.16.0
sip==4.19.21
six==1.15.0
sklearn==0.0
sniffio==1.2.0
soupsieve==2.1
spglib==1.16.0
Sphinx==1.8.5
-e git+https://github.com/ohebbi/PredictingSolidStateQubitCandidates.git@a2f06f0df9de4290ddc4c9e62191c23854145033#egg=src
ssh-import-id==5.10
stevedore==3.3.0
sympy==1.7.1
system-service==0.3
systemd-python==234
tabulate==0.8.7
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.1.0
toolz==0.11.1
tornado==6.1
tqdm==4.55.1
traitlets==5.0.5
twine==3.3.0
typing-extensions==3.7.4.3
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uncertainties==3.1.5
urllib3==1.26.2
usb-creator==0.3.7
virtualenv==20.4.2
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xgboost==1.3.3
xkit==0.0.0
xmltodict==0.12.0
zope.interface==4.7.1
